Support for multiple tabs in Resources#show action
This add support for display multiple fields on multiple tabs in the show action so we
can cleanup the view and isolate data.

Now you can specify the show action as follow :

  User:
    fields:
      default: full_name, address, phone, price_with_currency
      show:
        01_basic: email, created_at, updated_at, full_name, nationality
        02_billing: credit_card, credit_card_number
        03_address: street, city, country

The final result is convined with jQuery UI Tabs on the display. The old one-line field list
is also available.